+++ wire5/src/commonMain/kotlin/mqtt/wire5/control/packet/PublishMessage.kt
@@ -643,7 +643,7 @@ data class PublishMessage<ApplicationMessage : Any>(
             val variableSize = variableHeaderSized.first
             val deserializationParameters = DeserializationParameters(
                 buffer,
-                (remainingLength - variableSize).toUShort(),
+                remainingLength - variableSize,
                 variableHeader.topicName
             )
             val genericType = buffer.readGenericType(deserializationParameters)
